requirements.txt
google-api-python-client==2.108.0
kafka-python==2.0.2
redis==4.6.0
streamlit==1.28.1
transformers==4.35.2
torch==2.1.1
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.24.3
google-generativeai==0.3.2
streamz==0.6.4
plotly==5.17.0
asyncio==3.4.3
schedule==1.2.0
emoji==2.8.0
